feat: add PDF receipt form data type and formatting function
